% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{calc_dist}
\alias{calc_dist}
\title{Calculates isolate group membership}
\usage{
calc_dist(K, plot = TRUE)
}
\arguments{
\item{K}{The K value (or range) used to calculate ancestry matrix}

\item{plot}{TRUE/FALSE: Plots relationships between isolates visually}
}
\value{
membership.txt file containing membership assignments at each K value

Kplot.tiff TIFF file containing network plot at specified K value
}
\description{
This function calculates the euclidean distance from ancestry matrices (Q). It assigns isolates to groups using a pairwise distance of 0 (+/- 1 SD). These membership groups are then plotted (if plot = TRUE) as a network and output as a membership table at each specified K value.
}
\examples{
calc_dist(K = 4)
calc_dist(K = 2:20, plot = FALSE)
}
