% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_labels.R
\name{extract_labels}
\alias{extract_labels}
\title{Extracts isolate names from VCF file}
\usage{
extract_labels(file)
}
\arguments{
\item{file}{Unzipped VCF file}
}
\value{
List of isolate names
}
\description{
This function calculates the euclidean distance from ancestry matrices (Q). It assigns isolates to groups using a pairwise distance of 0 (+/- 1 SD). These membership groups are then plotted (if plot = TRUE) as a network and output as a membership table at each specified K value.
}
\examples{
extract_labels("example.vcf")
}
