% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_Q.R
\name{make_Q}
\alias{make_Q}
\title{Calculate ancestry coefficients}
\usage{
make_Q(K, CPU = 1, alpha = 10, tolerance = 1e-05, iterations = 200, ploidy = 2)
}
\arguments{
\item{K}{The K value (or range) used to calculate ancestry matrix}

\item{CPU}{A number of CPUs to run the parallel version of the algorithm (LEA)}

\item{alpha}{A numeric value corresponding to the snmf regularization parameter (LEA)}

\item{tolerance}{A numeric value for the tolerance error (LEA)}

\item{iterations}{An integer for the maximum number of iterations in SNMF algorithm (LEA)}

\item{ploidy}{1 if haploid, 2 if diploid, n if n-ploid}

\item{plot}{TRUE/FALSE: Plots relationships between isolates visually}
}
\value{
.Q file
}
\description{
This function creates individual ancestry coefficients (implemented through LEA) of subsetted .geno files. The subsetted .geno files are located in 'subsets/'.
}
\examples{
make_Q(K = 2:20, CPU = 1, alpha = 10, tolerance = 1e-05, iterations = 200, ploidy = 2)
}
