% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_subsets.R
\name{make_subsets}
\alias{make_subsets}
\title{Make subsets of .geno file for euclidean distance calculations}
\usage{
make_subsets(input.file, subsets = 100, snps = 1000)
}
\arguments{
\item{input.file}{.geno file created by LEA::vcf2geno()}

\item{subsets}{Number of ancestry calculations at each K value}

\item{snps}{Number of SNPs sub-sampled for ancestry calculations}
}
\value{
.geno file
}
\description{
This function creates multiple smaller .geno files from larger .geno file.
}
\examples{
make_subsets("example.geno", snps = 1000, subsets = 100)
}
