% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geno.R
\name{read.geno}
\alias{read.geno}
\title{Reads .geno file for make_subsets()}
\usage{
read.geno(input.file)
}
\arguments{
\item{input.file}{.geno file created by LEA::vcf2geno()}
}
\value{
.geno data frame
}
\description{
Reads .geno file for make_subsets()
}
